+++ crates/commands/src/lib.rs
@@ -2,6 +2,7 @@ mod build;
 mod change_env;
 mod clean;
 mod dep_graph;
+mod find_envs;
 mod fwd_search;
 mod placeholders;
 
@@ -10,5 +11,6 @@ pub use self::{
     change_env::{change_environment, ChangeEnvironmentResult},
     clean::{CleanCommand, CleanTarget},
     dep_graph::show_dependency_graph,
+    find_envs::find_environments,
     fwd_search::{ForwardSearch, ForwardSearchError},
 };